Basics of posting data done
The currently implemented structs are post-able to the api. A generic
`Submit` method has been implemented as have some handers for a more
"structured" way of posting data. Currently only very basic
clarification and submission has been implmented. Next steps
should/could be implementing more types to make the interaction more
useable.

Another change is the approach to JSON unmarshalling. Previously all
types implementing ApiType had to be pointer types, but this was anoying
when trying to submit an object. Due to this refactoring non-pointer
instances can be submitted.

The interaction itself also needs to be made more consistent. A few
different styles are now used for the same type of interaction which
looks sloppy.