Formatting times
We stuck with the default Duration and Time format so far, but to me these are a bit hard to read and verbose for humans. I propose we change the duration to only show h/m/s values when they are non-zero, and to change the date format to a simple local time and either the date or 'today'.

This affects any output with time (submissions, clarifications), but the contest output shows both. It would go from:
Contests (2):
   Id      Name                    Start Time                     Length  Status
   finals  ICPC World Finals 2019  2019-04-04 06:50:25 -0400 EDT  5h0m0s  Contest over
   comm    practice                2020-02-21 09:57:00 -0500 EST  2h0m0s  Contest over
to:
Contests (2):
   Id      Name                    Start Time                      Length  Status
   comm    practice                9:57:00am on February 21, 2020  2h      Contest over
   finals  ICPC World Finals 2019  4:51:56pm today                 5h      Scheduled